Add link fetching section of downloader (#7)
Refer to https://github.com/NASA-IMPACT/hls-sentinel2-downloader-serverless/pull/7 for a full rundown of what changed.

TL;DR: This PR introduces the Link Fetching components of the hls-sentinel2-downloader-serverless project. It handles Link Fetching for 21 days into the past, stores and tracks these links in an RDS instance in the granule table. The total available and processed (fetched) links are tracked in the granule_count and the last fetched link time is tracked in the status table. Links are then put into the to-download SQS queue so that they can be processed by further functions later on.