Add downloading section of downloader (#8)
This PR introduces the Downloader components of the hls-sentinel2-downloader-serverless project. It handles the downloading of Sentinel 2 Images into the UPLOAD_BUCKET on S3. It is run with a max concurrency of 15 which maps to the dedicated connections we have for IntHub. The granule table is updated throughout to signal that a download has started, to update the checksum value of the image, to mark it is downloaded or not and whether a retry has been added.